+++ game/src/main/scala/hexacraft/main/MainRouter.scala
@@ -1,11 +1,12 @@
 package hexacraft.main
 
-import hexacraft.game.*
-import hexacraft.game.Menus.JoinWorldChooserMenu
+import hexacraft.client.BlockTextureLoader
+import hexacraft.game.GameKeyboard
 import hexacraft.gui.Scene
 import hexacraft.infra.audio.AudioSystem
 import hexacraft.infra.fs.FileSystem
 import hexacraft.infra.window.CursorMode
+import hexacraft.server.WorldProviderFromFile
 import hexacraft.util.Channel
 import hexacraft.world.WorldSettings
