+++ game/src/main/scala/hexacraft/main/Menus.scala
@@ -1,13 +1,12 @@
-package hexacraft.game
+package hexacraft.main
 
+import com.martomate.nbt.Nbt
 import hexacraft.gui.{LocationInfo, RenderContext, Scene}
-import hexacraft.gui.comp.{Button, Component, GUITransformation, Label, ScrollPane, SubComponents, TextField}
+import hexacraft.gui.comp.*
 import hexacraft.infra.fs.{FileSystem, NbtIO}
 import hexacraft.renderer.TextureSingle
 import hexacraft.util.Channel
 import hexacraft.world.WorldSettings
-
-import com.martomate.nbt.Nbt
 
 import java.io.File
 import java.nio.file.Path
