+++ game/src/main/scala/hexacraft/server/GameServer.scala
@@ -1,9 +1,8 @@
 package hexacraft.server
 
-import hexacraft.game.{GameKeyboard, NetworkPacket, PlayerInputHandler, PlayerPhysicsHandler, Ray, RayTracer}
+import hexacraft.game.{GameKeyboard, NetworkPacket, PlayerInputHandler, PlayerPhysicsHandler}
 import hexacraft.util.Result
 import hexacraft.world.*
-import hexacraft.world.EntityEvent.Despawned
 import hexacraft.world.block.{Block, BlockState}
 import hexacraft.world.chunk.ChunkColumnData
 import hexacraft.world.coord.*
