+++ game/src/test/scala/hexacraft/world/RayTracerTest.scala
@@ -1,9 +1,8 @@
-package hexacraft.game
+package hexacraft.world
 
 import hexacraft.world.*
 import hexacraft.world.block.{Block, BlockState}
 import hexacraft.world.coord.{BlockCoords, BlockRelWorld}
-
 import munit.FunSuite
 import org.joml.Vector2f
